# api/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the Click Proposta v1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/verify-phone": {
      "get": {
        "operationId": "getUserByPhone",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "phone",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "phone": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string",
                          "format": "email",
                          "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                        },
                        "avatarUrl": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "nullable": true,
                              "enum": [
                                null
                              ]
                            }
                          ]
                        },
                        "cnpj": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "nullable": true,
                              "enum": [
                                null
                              ]
                            }
                          ]
                        },
                        "address": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "nullable": true,
                              "enum": [
                                null
                              ]
                            }
                          ]
                        },
                        "planType": {
                          "type": "string"
                        },
                        "planExpiresAt": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "nullable": true,
                              "enum": [
                                null
                              ]
                            }
                          ]
                        },
                        "countProposalsInMonth": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "id",
                        "phone",
                        "email",
                        "avatarUrl",
                        "cnpj",
                        "address",
                        "planType",
                        "planExpiresAt",
                        "countProposalsInMonth"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "user"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/payment": {
      "post": {
        "operationId": "createNewPayment",
        "tags": [
          "Payments"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "customer": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "minLength": 3
                      },
                      "email": {
                        "type": "string",
                        "format": "email",
                        "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                      },
                      "cellphone": {
                        "type": "string",
                        "minLength": 10,
                        "maxLength": 11
                      },
                      "cpf": {
                        "type": "string",
                        "minLength": 11,
                        "maxLength": 14
                      }
                    },
                    "required": [
                      "name",
                      "email",
                      "cellphone",
                      "cpf"
                    ]
                  }
                },
                "required": [
                  "customer"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}